Glob: rework how the path results are retrieved.
Use a CCode attribute to associate the path count member with the paths member.
Remove the 'length' property of VFS.Glob to reflect this.
Change the 'paths' property to a custom getter function, so that the path count
is properly returned as well.
Update the test and Config Schema to reflect the changed VFS.Glob API.