Fix the keyfile configuration backend.
Gconf is deprecated and the keyfile backend is broken. Fix several infinite
loops and segmentation faults found while testing. Some code was rewritten
to minimize the File I/O operations. Each applet generates a FileMonitor for
avant-window-navigator.ini and a single change there will trigger a cascade of
checks which currently resulted in a ton of I/O operations and in the worst case
infinite loops.